% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{addSets}
\alias{addSets}
\title{Add set numbers to introns}
\usage{
addSets(clusterGRanges.noset)
}
\arguments{
\item{clusterGRanges.noset}{Granges object with a cluster of intron locations}
}
\value{
Granges object with a cluster of intron locations and corresponding set numbers
}
\description{
Converts a group of introns into all non-overlapping sets
}
\author{
Beth Signal
}
\keyword{internal}
