% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafCutter.R
\name{alternativeIntronUsage}
\alias{alternativeIntronUsage}
\title{Create transcripts with alternative intron usage}
\usage{
alternativeIntronUsage(
  altIntronLocs,
  exons,
  replaceInternalExons = TRUE,
  junctions = NULL
)
}
\arguments{
\item{altIntronLocs}{data.frame containing information from the
per_intron_results.tab file output from leafcutter.
Note that only one cluster of alternative introns can be processed at a time.}

\item{exons}{GRanges object made from a GTF with ONLY exon annotations
(no gene, transcript, CDS etc.)}

\item{replaceInternalExons}{replace any internal transcript exons with inferred exons from multi-intron leafcutter intron sets?}

\item{junctions}{GRanges object from readLeafcutterJunctions()}
}
\value{
GRanges object with all potential alternative isoforms skipping the
introns specified in either the upregulated or downregulated locations
}
\description{
Creates transcript isoforms from alternative intron usage tested by leafcutter
}
\examples{
leafcutterFiles <- list.files(system.file("extdata", "leaf_small/",
    package = "GeneStructureTools"
), full.names = TRUE)
leafcutterIntrons <- read.delim(leafcutterFiles[grep(
    "intron_results",
    leafcutterFiles
)], stringsAsFactors = FALSE)
gtf <- rtracklayer::import(system.file("extdata", "gencode.vM25.small.gtf",
    package = "GeneStructureTools"
))
exons <- gtf[gtf$type == "exon"]
# single cluster processing
cluster <- leafcutterIntrons[leafcutterIntrons$cluster == "chr17:clu_20975_+", ]
altIsoforms20975 <- alternativeIntronUsage(cluster, exons)
unique(altIsoforms20975$transcript_id)
}
\seealso{
Other leafcutter splicing isoform creation: 
\code{\link{leafcutterIntronsToExons}()},
\code{\link{makeNewLeafExonsUnanchored}()},
\code{\link{makeNewLeafExons}()},
\code{\link{readLeafcutterJunctions}()}
}
\author{
Beth Signal
}
\concept{leafcutter splicing isoform creation}
