% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altTransStartEnd.R
\name{alterTranscriptStartEnds}
\alias{alterTranscriptStartEnds}
\title{Find transcripts overlapping first/last exons and alter their transcription start site or transcription end site}
\usage{
alterTranscriptStartEnds(whippetDataSet, exons, unfilteredWDS, type = NA)
}
\arguments{
\item{whippetDataSet}{whippetDataSet generated from \code{readWhippetDataSet()}}

\item{exons}{GRanges object made from a GTF containing exon coordinates}

\item{unfilteredWDS}{unfiltered whippetDataSet generated from \code{readWhippetDataSet()}
Note that this should contain ALL TS/TE events, regardless of significance, as these are used to contruct the entire exon range.}

\item{type}{type of Whippet event (TS/TE).
Note only one event type should be processed at a time.}
}
\value{
GRanges object with transcripts containing alternative junctions.
}
\description{
Find transcripts overlapping first/last exons and alter their transcription start site or transcription end site
}
\examples{
gtf <- rtracklayer::import(system.file("extdata", "gencode.vM25.small.gtf", package = "GeneStructureTools"))
exons <- gtf[gtf$type == "exon"]
g <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10

whippetFiles <- system.file("extdata", "whippet_small/",
    package = "GeneStructureTools"
)
wds <- readWhippetDataSet(whippetFiles)

wds.TS <- filterWhippetEvents(wds, eventTypes = "TS", psiDelta = 0.1, probability = 0.95)
alterTranscriptStartEnds(whippetDataSet = wds.TS, exons = exons, unfilteredWDS = wds, type = "TS")
}
\seealso{
Other whippet splicing isoform creation: 
\code{\link{addIntronInTranscript}()},
\code{\link{findExonContainingTranscripts}()},
\code{\link{findIntronContainingTranscripts}()},
\code{\link{findJunctionPairs}()},
\code{\link{replaceJunction}()},
\code{\link{skipExonInTranscript}()}
}
\author{
Beth Signal
}
\concept{whippet splicing isoform creation}
