% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altSplicingrMATS.R
\name{altIntronRmats}
\alias{altIntronRmats}
\title{Generate isoforms with and without a retain intron from RMATS data
shortcut function: uses the same base function for intron retention}
\usage{
altIntronRmats(rmatsEvents, exons)
}
\arguments{
\item{rmatsEvents}{data.frame containing RMATS RI events}

\item{exons}{reference exons GRanges}
}
\value{
GRanges retained and skipped intron isoforms
}
\description{
Generate isoforms with and without a retain intron from RMATS data
shortcut function: uses the same base function for intron retention
}
\examples{
gtf <- rtracklayer::import(system.file("extdata", "gencode.vM25.small.gtf", package = "GeneStructureTools"))
exons <- gtf[gtf$type == "exon"]
g <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10

rmats_directory <- system.file("extdata", "rmats_small/", package = "GeneStructureTools")
rds <- readRmatsDataSet(rmats_directory)
rds.filtered <- filterRmatsEvents(rds, FDR = 0.01, psiDelta = 0.1)

diffSplice.RI <- extractEvent(rds.filtered, "RI")
isoforms.RI <- altIntronRmats(diffSplice.RI, exons = exons)
}
\seealso{
Other rmats data processing: 
\code{\link{altSpliceSiteRmats}()},
\code{\link{annotateEventCoords}()},
\code{\link{annotateOverlapRmats}()},
\code{\link{betweenNumbers}()},
\code{\link{duplicateReference}()},
\code{\link{exonsToIntrons}()},
\code{\link{extractEvent}()},
\code{\link{filterRmatsEvents}()},
\code{\link{readRmatsDataSet}()},
\code{\link{reformatExons}()},
\code{\link{removeDuplicatePairs}()},
\code{\link{removeExonsBetween}()},
\code{\link{rmatsTranscriptChangeSummary}()},
\code{\link{skipExonByJunction}()},
\code{\link{splitLongExons}()}
}
\author{
Beth Signal
}
\concept{rmats data processing}
