% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altSplicingrMATS.R
\name{altSpliceSiteRmats}
\alias{altSpliceSiteRmats}
\title{Generate isoforms with different 5' or 3' splice site usage from RMATS data}
\usage{
altSpliceSiteRmats(rmatsEvents, exons, eventType)
}
\arguments{
\item{rmatsEvents}{data.frame containing RMATS RI events}

\item{exons}{reference exons GRanges}

\item{eventType}{type of event. "A5E" - alternative 5', or "A3E" - alternative 3'}
}
\value{
GRanges of isoforms
}
\description{
Generate isoforms with different 5' or 3' splice site usage from RMATS data
}
\examples{
gtf <- rtracklayer::import(system.file("extdata", "gencode.vM25.small.gtf", package = "GeneStructureTools"))
exons <- gtf[gtf$type == "exon"]
g <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10

rmats_directory <- system.file("extdata", "rmats_small/", package = "GeneStructureTools")
rds <- readRmatsDataSet(rmats_directory)
rds.filtered <- filterRmatsEvents(rds, FDR = 0.01, psiDelta = 0.1)

diffSplice.A5E <- extractEvent(rds.filtered, "A5SS")
isoforms.A5E <- altSpliceSiteRmats(diffSplice.A5E, exons = exons, eventType = "A5SS")

diffSplice.A3E <- extractEvent(rds.filtered, "A3SS")
isoforms.A3E <- altSpliceSiteRmats(diffSplice.A3E, exons = exons, eventType = "A3SS")
}
\seealso{
Other rmats data processing: 
\code{\link{altIntronRmats}()},
\code{\link{annotateEventCoords}()},
\code{\link{annotateOverlapRmats}()},
\code{\link{betweenNumbers}()},
\code{\link{duplicateReference}()},
\code{\link{exonsToIntrons}()},
\code{\link{extractEvent}()},
\code{\link{filterRmatsEvents}()},
\code{\link{readRmatsDataSet}()},
\code{\link{reformatExons}()},
\code{\link{removeDuplicatePairs}()},
\code{\link{removeExonsBetween}()},
\code{\link{rmatsTranscriptChangeSummary}()},
\code{\link{skipExonByJunction}()},
\code{\link{splitLongExons}()}
}
\author{
Beth Signal
}
\concept{rmats data processing}
