% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{annotateEventCoords}
\alias{annotateEventCoords}
\title{annotate event coordinates with exon/transcript/gene ids.}
\usage{
annotateEventCoords(eventCoords, betweenExons, exons)
}
\arguments{
\item{eventCoords}{Granges with event coordinates}

\item{betweenExons}{data.frame with related differential splicing event ids and reference transcript_ids}

\item{exons}{reference exons GRanges}
}
\value{
eventCoords Granges with annotations
}
\description{
annotate event coordinates with exon/transcript/gene ids.
}
\seealso{
Other rmats data processing: 
\code{\link{altIntronRmats}()},
\code{\link{altSpliceSiteRmats}()},
\code{\link{annotateOverlapRmats}()},
\code{\link{betweenNumbers}()},
\code{\link{duplicateReference}()},
\code{\link{exonsToIntrons}()},
\code{\link{extractEvent}()},
\code{\link{filterRmatsEvents}()},
\code{\link{readRmatsDataSet}()},
\code{\link{reformatExons}()},
\code{\link{removeDuplicatePairs}()},
\code{\link{removeExonsBetween}()},
\code{\link{rmatsTranscriptChangeSummary}()},
\code{\link{skipExonByJunction}()},
\code{\link{splitLongExons}()}
}
\author{
Beth Signal
}
\concept{rmats data processing}
\keyword{internal}
