% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGvisTracks.R
\name{annotateGeneModel}
\alias{annotateGeneModel}
\title{Annotate a GRanges gene model with ORF boundries for visualisation with Gviz}
\usage{
annotateGeneModel(transcripts, orfs)
}
\arguments{
\item{transcripts}{GRanges of gene model to be visualised}

\item{orfs}{ORF predictions. Created by getORFs()}
}
\value{
data.frame of a gene model for visualisation
}
\description{
Annotate a GRanges gene model with ORF boundries for visualisation with Gviz
}
\author{
Beth Signal
}
