% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{annotateOverlapRmats}
\alias{annotateOverlapRmats}
\title{Find overlaps from a rmats GRanges to reference exons and annotate with ids}
\usage{
annotateOverlapRmats(rmatsGRanges, exons, exon_number = 1)
}
\arguments{
\item{rmatsGRanges}{rmats event GRanges object}

\item{exons}{reference exons GRanges}

\item{exon_number}{which exon in the rmats event this is for (only used for annotating output)}
}
\value{
data.frame with overlapping event/exons
}
\description{
Find overlaps from a rmats GRanges to reference exons and annotate with ids
}
\seealso{
Other rmats data processing: 
\code{\link{altIntronRmats}()},
\code{\link{altSpliceSiteRmats}()},
\code{\link{annotateEventCoords}()},
\code{\link{betweenNumbers}()},
\code{\link{duplicateReference}()},
\code{\link{exonsToIntrons}()},
\code{\link{extractEvent}()},
\code{\link{filterRmatsEvents}()},
\code{\link{readRmatsDataSet}()},
\code{\link{reformatExons}()},
\code{\link{removeDuplicatePairs}()},
\code{\link{removeExonsBetween}()},
\code{\link{rmatsTranscriptChangeSummary}()},
\code{\link{skipExonByJunction}()},
\code{\link{splitLongExons}()}
}
\author{
Beth Signal
}
\concept{rmats data processing}
\keyword{internal}
