% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{betweenNumbers}
\alias{betweenNumbers}
\title{Generate vector of integers between two numbers (non-inclusive)}
\usage{
betweenNumbers(a, b)
}
\arguments{
\item{a}{first integer}

\item{b}{second integer}
}
\value{
vector of integers
}
\description{
Generate vector of integers between two numbers (non-inclusive)
}
\seealso{
Other rmats data processing: 
\code{\link{altIntronRmats}()},
\code{\link{altSpliceSiteRmats}()},
\code{\link{annotateEventCoords}()},
\code{\link{annotateOverlapRmats}()},
\code{\link{duplicateReference}()},
\code{\link{exonsToIntrons}()},
\code{\link{extractEvent}()},
\code{\link{filterRmatsEvents}()},
\code{\link{readRmatsDataSet}()},
\code{\link{reformatExons}()},
\code{\link{removeDuplicatePairs}()},
\code{\link{removeExonsBetween}()},
\code{\link{rmatsTranscriptChangeSummary}()},
\code{\link{skipExonByJunction}()},
\code{\link{splitLongExons}()}
}
\author{
Beth Signal
}
\concept{rmats data processing}
\keyword{internal}
