% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTranscripts.R
\name{compareContains}
\alias{compareContains}
\title{compare two strings of concatenated sequence features}
\usage{
compareContains(contains)
}
\arguments{
\item{contains}{orf vector of two concatenated sequence features strings}
}
\value{
data.frame with orf changes
}
\description{
compare two strings of concatenated sequence features
}
\author{
Beth Signal
}
\keyword{internal}
