% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{coordinates}
\alias{coordinates}
\alias{coordinates,whippetDataSet-method}
\title{Method coordinates}
\usage{
coordinates(whippetDataSet)

\S4method{coordinates}{whippetDataSet}(whippetDataSet)
}
\arguments{
\item{whippetDataSet}{whippetDataSet generated from \code{readWhippetDataSet()}}
}
\value{
whippet splicing event coordinates as a GRanges object
}
\description{
Method coordinates
}
\examples{
whippetFiles <- system.file("extdata","whippet/",
package = "GeneStructureTools")
wds <- readWhippetDataSet(whippetFiles)

coordinates <- coordinates(wds)
}
