% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longestORF.R
\name{cumsumANDpad}
\alias{cumsumANDpad}
\title{Cumulative sum of a sequence of numbers, padded with NA}
\usage{
cumsumANDpad(x, padLength)
}
\arguments{
\item{x}{input numeric vector}

\item{padLength}{length to pad output to}
}
\value{
vector with cumulative sum, padded with NA
}
\description{
Cumulative sum of a sequence of numbers, padded with NA
}
\examples{
x <- c(1,4,7,2,5)
cumsumANDpad(x, 10)
}
\author{
Beth Signal
}
