% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{diffSplicingResults}
\alias{diffSplicingResults}
\alias{diffSplicingResults,whippetDataSet-method}
\title{Method diffSplicingResults}
\usage{
diffSplicingResults(whippetDataSet)

\S4method{diffSplicingResults}{whippetDataSet}(whippetDataSet)
}
\arguments{
\item{whippetDataSet}{FUUUUUUCk}
}
\description{
Method diffSplicingResults
}
