% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntronRetention.R
\name{distance_to_junction_b}
\alias{distance_to_junction_b}
\title{get distance to second exon-exon junction}
\usage{
distance_to_junction_b(diffs)
}
\arguments{
\item{diffs}{distances to junctions}
}
\value{
distance to closest second exon-exon junction
}
\description{
get distance to second exon-exon junction
}
\author{
Beth Signal
}
