% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{duplicateReference}
\alias{duplicateReference}
\title{Duplicate a reference Granges (exon-level) for each diff-splicing event/transcript combination required.}
\usage{
duplicateReference(betweenExons, exons)
}
\arguments{
\item{betweenExons}{data.frame with related differential splicing event ids and reference transcript_ids}

\item{exons}{reference exons GRanges}
}
\value{
Granges with transcripts duplicated
}
\description{
Duplicate a reference Granges (exon-level) for each diff-splicing event/transcript combination required.
}
\seealso{
Other rmats data processing: 
\code{\link{altIntronRmats}()},
\code{\link{altSpliceSiteRmats}()},
\code{\link{annotateEventCoords}()},
\code{\link{annotateOverlapRmats}()},
\code{\link{betweenNumbers}()},
\code{\link{exonsToIntrons}()},
\code{\link{extractEvent}()},
\code{\link{filterRmatsEvents}()},
\code{\link{readRmatsDataSet}()},
\code{\link{reformatExons}()},
\code{\link{removeDuplicatePairs}()},
\code{\link{removeExonsBetween}()},
\code{\link{rmatsTranscriptChangeSummary}()},
\code{\link{skipExonByJunction}()},
\code{\link{splitLongExons}()}
}
\author{
Beth Signal
}
\concept{rmats data processing}
\keyword{internal}
