% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exonReformatting.R
\name{exonsToIntrons}
\alias{exonsToIntrons}
\title{Generate a introns Granges from an exons Granges}
\usage{
exonsToIntrons(exons)
}
\arguments{
\item{exons}{reference exons GRanges}
}
\value{
reference introns GRanges
}
\description{
Generate a introns Granges from an exons Granges
}
\examples{
gtf <- rtracklayer::import(system.file("extdata", "gencode.vM25.small.gtf",
    package = "GeneStructureTools"
))
exons <- gtf[gtf$type == "exon"]
introns <- exonsToIntrons(exons)
}
\seealso{
Other rmats data processing: 
\code{\link{altIntronRmats}()},
\code{\link{altSpliceSiteRmats}()},
\code{\link{annotateEventCoords}()},
\code{\link{annotateOverlapRmats}()},
\code{\link{betweenNumbers}()},
\code{\link{duplicateReference}()},
\code{\link{extractEvent}()},
\code{\link{filterRmatsEvents}()},
\code{\link{readRmatsDataSet}()},
\code{\link{reformatExons}()},
\code{\link{removeDuplicatePairs}()},
\code{\link{removeExonsBetween}()},
\code{\link{rmatsTranscriptChangeSummary}()},
\code{\link{skipExonByJunction}()},
\code{\link{splitLongExons}()}
}
\author{
Beth Signal
}
\concept{rmats data processing}
