% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafCutter.R
\name{exonsToTranscripts}
\alias{exonsToTranscripts}
\title{Convert an exon-level gtf annotation to a transcript-level gtf annotation}
\usage{
exonsToTranscripts(gtf.exons)
}
\arguments{
\item{gtf.exons}{GRanges object with exons}
}
\value{
GRanges object with transcripts
}
\description{
Convert an exon-level gtf annotation to a transcript-level gtf annotation
}
\examples{
gtf <- rtracklayer::import(system.file("extdata","example_gtf.gtf",
package = "GeneStructureTools"))
gtf.exons <- gtf[gtf$type=="exon" & gtf$transcript_id=="ENSMUST00000126412.1"]
gtf.exons
gtf.transcripts <- exonsToTranscripts(gtf.exons)
gtf.transcripts
}
\author{
Beth Signal
}
