% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{extractEvent}
\alias{extractEvent}
\alias{extractEvent,rmatsDataSet-method}
\title{Method extractEvent}
\usage{
extractEvent(rmatsDataSet, eventType)

\S4method{extractEvent}{rmatsDataSet}(rmatsDataSet, eventType)
}
\arguments{
\item{rmatsDataSet}{rmatsDataSet generated from \code{readRmatsDataSet()}}

\item{eventType}{specific event type to extract results for.
Must be SE/MXE/RI/A5SS/A3SS.}
}
\value{
differential splicing results data.frame
(originally from a whippet .diff file)
}
\description{
Method extractEvent
}
\examples{
rmats_directory <- system.file("extdata", "rmats_small/",
    package = "GeneStructureTools"
)
rds <- readRmatsDataSet(rmats_directory)
rds.SE <- extractEvent(rds, "SE")
}
\seealso{
Other rmats data processing: 
\code{\link{altIntronRmats}()},
\code{\link{altSpliceSiteRmats}()},
\code{\link{annotateEventCoords}()},
\code{\link{annotateOverlapRmats}()},
\code{\link{betweenNumbers}()},
\code{\link{duplicateReference}()},
\code{\link{exonsToIntrons}()},
\code{\link{filterRmatsEvents}()},
\code{\link{readRmatsDataSet}()},
\code{\link{reformatExons}()},
\code{\link{removeDuplicatePairs}()},
\code{\link{removeExonsBetween}()},
\code{\link{rmatsTranscriptChangeSummary}()},
\code{\link{skipExonByJunction}()},
\code{\link{splitLongExons}()}
}
\concept{rmats data processing}
