% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTranscripts.R
\name{featureTypes}
\alias{featureTypes}
\title{Evaluate changes to ORFs caused by alternative splicing}
\usage{
featureTypes(seqFeatures, orf)
}
\arguments{
\item{seqFeatures}{data.frame of uniprot sequence features}

\item{orf}{vector with open reading frame (amino acid sequence) string and gene_id}
}
\value{
data.frame with orf changes
}
\description{
Evaluate changes to ORFs caused by alternative splicing
}
\author{
Beth Signal
}
\keyword{internal}
