% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickIrfinder.R
\name{filterIrfEvents}
\alias{filterIrfEvents}
\title{Filter out significant events from a irfDataSet}
\usage{
filterIrfEvents(irfDataSet, FDR = 0.05, psiDelta = 0.1, idList = NA)
}
\arguments{
\item{irfDataSet}{irfDataSet generated from \code{readIrfDataSet()}}

\item{FDR}{maximum FDR required to call event as significant}

\item{psiDelta}{minimum change in psi required to call an event as significant}

\item{idList}{(optional) list of gene ids to filter for}
}
\value{
filtered irfDataSet
}
\description{
Filter out significant events from a irfDataSet
}
\examples{
irfinder_file <- list.files(system.file("extdata", "irf_small/", package = "GeneStructureTools"), full.names = TRUE)
irf <- readIrfDataSet(irfinder_file)
irf.filtered <- filterIrfEvents(irf, FDR = 0.01, psiDelta = 0.1)

# filter by gene id/name
irf.filtered <- filterIrfEvents(irf, idList = "Tmem208")
}
\seealso{
Other IRFinder data processing: 
\code{\link{irfTranscriptChangeSummary}()},
\code{\link{readIrfDataSet}()}
}
\author{
Beth Signal
}
\concept{IRFinder data processing}
