% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickRmats.R
\name{filterRmatsEvents}
\alias{filterRmatsEvents}
\title{Filter out significant events from a RMATS dataset}
\usage{
filterRmatsEvents(
  rmatsDataSet,
  FDR = 0.05,
  psiDelta = 0.1,
  idList = NA,
  minCounts = NA,
  medianCounts = NA,
  sampleTable
)
}
\arguments{
\item{rmatsDataSet}{rmatsDataSet generated from \code{readRmatsDataSet()}}

\item{FDR}{maximum FDR required to call event as significant}

\item{psiDelta}{minimum change in psi required to call an event as significant}

\item{idList}{(optional) list of gene ids to filter for}

\item{minCounts}{minumum number of counts for all replicates
in at least one condition to call an event as significant}

\item{medianCounts}{median count for all replicates
in at least one condition to call an event as significant}

\item{sampleTable}{data.frame with sample names and conditions.
Only needed if filtering with counts.}
}
\value{
filtered rmatsDataSet
}
\description{
Filter out significant events from a RMATS dataset
}
\examples{
rmats_directory <- system.file("extdata", "rmats_small/", package = "GeneStructureTools")
rds <- readRmatsDataSet(rmats_directory)
rds.filtered <- filterRmatsEvents(rds, FDR = 0.01, psiDelta = 0.1)
# filter by gene name/id
rds.Tmem208 <- filterRmatsEvents(rds, idList = "Tmem208", FDR = 1, psiDelta = 0)
}
\seealso{
Other rmats data processing: 
\code{\link{altIntronRmats}()},
\code{\link{altSpliceSiteRmats}()},
\code{\link{annotateEventCoords}()},
\code{\link{annotateOverlapRmats}()},
\code{\link{betweenNumbers}()},
\code{\link{duplicateReference}()},
\code{\link{exonsToIntrons}()},
\code{\link{extractEvent}()},
\code{\link{readRmatsDataSet}()},
\code{\link{reformatExons}()},
\code{\link{removeDuplicatePairs}()},
\code{\link{removeExonsBetween}()},
\code{\link{rmatsTranscriptChangeSummary}()},
\code{\link{skipExonByJunction}()},
\code{\link{splitLongExons}()}
}
\author{
Beth Signal
}
\concept{rmats data processing}
