% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickAnalysis.R
\name{filterSignificantWhippetEvents}
\alias{filterSignificantWhippetEvents}
\title{Filter out significant events from a whippet diff comparison}
\usage{
filterSignificantWhippetEvents(whippetDiff, probability = 0.95,
  psiDelta = 0.1, eventTypes = "all", whippetCounts = NULL,
  minCounts = NA, sampleTable)
}
\arguments{
\item{whippetDiff}{data.frame containing whippet diff data.
May be generated by readWhippetDIFFfiles()}

\item{probability}{minimum probability required to call event as significant}

\item{psiDelta}{minimum change in psi required to call an event as significant}

\item{eventTypes}{which event type to filter for? default = "all"}

\item{whippetCounts}{optional whippet counts data.frame for read count filtering.
May be generated by readWhippetPSIfiles()}

\item{minCounts}{minumum number of counts for all replicates
in at least one condition to call an event as significant}

\item{sampleTable}{data.frame with sample names and conditions.
Only needed if filtering with counts.}
}
\value{
filtered whippet differential comparison data.frame
}
\description{
Filter out significant events from a whippet diff comparison
}
\examples{
whippetFiles <- list.files(system.file("extdata","whippet/",
package = "GeneStructureTools"), full.names = TRUE)
diffFiles <- whippetFiles[grep(".diff", whippetFiles)]
whippetDiffSplice <- readWhippetDIFFfiles(diffFiles)
whippetDiffSplice.significant <- filterSignificantWhippetEvents(whippetDiffSplice)
}
\author{
Beth Signal
}
