% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverlapFunctions.R
\name{findDEXexonType}
\alias{findDEXexonType}
\title{Find a DEXSeq exons' biotype}
\usage{
findDEXexonType(DEX_exon_id, dex_gtf, gtf, set = "overlap")
}
\arguments{
\item{DEX_exon_id}{vector of DEXSeq exon ids}

\item{dex_gtf}{GRanges object of the DEXSeq formatted gtf}

\item{gtf}{GRanges object of the GTF annotated with exon biotypes - i.e. exon, CDS, UTR}

\item{set}{which overlapping set of exon biotypes to return - to, from, and/or overlap}
}
\value{
overlaping types
}
\description{
Find a DEXSeq exons' biotype
}
\examples{
gtf_file <- system.file("extdata","gencode.vM14.annotation.small.gtf",
package = "GeneStructureTools")
dex_gtf_file <- system.file("extdata","gencode.vM14.annotation.dexseq.small.gtf",
package = "GeneStructureTools")

gtf <- rtracklayer::import(gtf_file)
gtf <- UTR2UTR53(gtf)
dex_gtf <- rtracklayer::import(dex_gtf_file)

findDEXexonType("ENSMUSG00000032366.15:E028", dex_gtf, gtf)

dex_results_file <- system.file("extdata","dexseq_results.small.txt",
package = "GeneStructureTools")
dex_results <- read.table(dex_results_file, sep="\\t")

findDEXexonType(rownames(dex_results), dex_gtf, gtf)

}
\author{
Beth Signal
}
