% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntronRetention.R
\name{findIntronContainingTranscripts}
\alias{findIntronContainingTranscripts}
\title{Given the location of a whole retained intron, find transcripts which splice out this intron}
\usage{
findIntronContainingTranscripts(whippetDataSet, gtf.exons, match = "exact")
}
\arguments{
\item{whippetDataSet}{whippetDataSet generated from \code{readWhippetDataSet()}}

\item{gtf.exons}{GRanges object made from a GTF with ONLY exon annotations
(no gene, transcript, CDS etc.)}

\item{match}{what type of matching to perform? exact = only exons which bound the intron exactly,
introns = any exon pairs which overlap the intron,
all = any exon pairs AND single exons which overlap the intron}
}
\value{
data.frame with all flanking exon pairs
}
\description{
Given the location of a whole retained intron, find transcripts which splice out this intron
}
\examples{
whippetFiles <- system.file("extdata","whippet/",
package = "GeneStructureTools")
wds <- readWhippetDataSet(whippetFiles)
wds <- filterWhippetEvents(wds)

gtf <- rtracklayer::import(system.file("extdata","example_gtf.gtf",
package = "GeneStructureTools"))
gtf.exons <- gtf[gtf$type=="exon"]
g <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10

wds.intronRetention <- filterWhippetEvents(wds, eventTypes="RI")
exons.intronRetention <- findIntronContainingTranscripts(wds.intronRetention, gtf.exons)
}
\author{
Beth Signal
}
