% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AltSplicingJunctionSupported.R
\name{findJunctionPairs}
\alias{findJunctionPairs}
\title{Find alternative junctions for Whippet alternative splicing events}
\usage{
findJunctionPairs(eventCoords, jncCoords, type = NA)
}
\arguments{
\item{eventCoords}{GRanges object with Whippet event coordinates}

\item{jncCoords}{GRanges object with Whippet junctions. Generated by formatJunctions()}

\item{type}{type of Whippet event (AA/AD/AF/AL).
Note only one event type should be processed at a time.}
}
\value{
GRanges object with alternative junctions.
Each event should have a set of X (for which the psi measurement is reported) junctions,
and alternative Y junctions.
}
\description{
Find junctions that pair with each end of an AA (alt. acceptor) or
AD (alt. donor) whippet range
Find junctions that pair with the upsteam/downstream exon of an
AF (alt. first exon) or an AL (alt. last exon)
}
\examples{
whippetFiles <- list.files(system.file("extdata","whippet/",
package = "GeneStructureTools"), full.names = TRUE)
jncFiles <- whippetFiles[grep(".jnc", whippetFiles)]
whippetJNC <- readWhippetJNCfiles(jncFiles)
junctionCoords <- formatJunctions(whippetJNC)
diffFiles <- whippetFiles[grep(".diff", whippetFiles)]
whippetDiffSplice <- readWhippetDIFFfiles(diffFiles)
whippetCoords <- formatWhippetEvents(whippetDiffSplice)

event.altAce <- whippetDiffSplice[which(whippetDiffSplice$type=="AA")[1],]
coords.altAce <- whippetCoords[whippetCoords$id \%in\% event.altAce$coord]
jncPairs.altAce <- findJunctionPairs(coords.altAce,junctionCoords, type="AA")

event.altDon <- whippetDiffSplice[which(whippetDiffSplice$type=="AD")[1],]
coords.altDon <- whippetCoords[whippetCoords$id \%in\% event.altDon$coord]
jncPairs.altDon <- findJunctionPairs(coords.altDon,junctionCoords, type="AD")

event.altFirst <- whippetDiffSplice[which(whippetDiffSplice$type=="AF")[1],]
coords.altFirst <- whippetCoords[whippetCoords$id \%in\% event.altFirst$coord]
jncPairs.altFirst <- findJunctionPairs(coords.altFirst,junctionCoords, type="AF")

event.altLast <- whippetDiffSplice[which(whippetDiffSplice$type=="AL")[1],]
coords.altLast <- whippetCoords[whippetCoords$id \%in\% event.altLast$coord]
jncPairs.altLast <- findJunctionPairs(coords.altLast,junctionCoords, type="AL")
}
\author{
Beth Signal
}
