% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{findOverlaps.junc}
\alias{findOverlaps.junc}
\title{Find overlaps where the start/end coordinates are the same}
\usage{
findOverlaps.junc(query, subject, type = c("start", "end"))
}
\arguments{
\item{query}{a GRanges object}

\item{subject}{a GRanges object}
}
\value{
Hits object
}
\description{
Find overlaps where the start/end coordinates are the same
}
\author{
Beth Signal
}
\concept{data processing}
\keyword{internal}
