% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AltSplicingJunctionSupported.R
\name{formatJunctions}
\alias{formatJunctions}
\title{Format Whippet junctions as a GRanges object}
\usage{
formatJunctions(junctions)
}
\arguments{
\item{junctions}{data.frame containing junction location information.
May be generated by readWhippetJNCfiles()}
}
\value{
GRanges object with junctions
}
\description{
Format Whippet junctions as a GRanges object
}
\examples{
whippetFiles <- list.files(system.file("extdata","whippet/",
 package = "GeneStructureTools"), full.names = TRUE)
jncFiles <- whippetFiles[grep(".jnc", whippetFiles)]
whippetJNC <- readWhippetJNCfiles(jncFiles)
junctionCoords <- formatJunctions(whippetJNC)
}
\author{
Beth Signal
}
