% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWhippetFiles.R
\name{formatWhippetEvents}
\alias{formatWhippetEvents}
\title{Format Whippet co-ordinates as a GRanges object}
\usage{
formatWhippetEvents(whippet)
}
\arguments{
\item{whippet}{data.frame containing event location information. May be generated by readWhippetDIFFfiles()}
}
\value{
GRanges object with events
}
\description{
Format Whippet co-ordinates as a GRanges object
}
\author{
Beth Signal
}
