% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntronRetention.R
\name{get_orfs}
\alias{get_orfs}
\title{Get open reading frames for transcripts}
\usage{
get_orfs(transcripts, BSgenome = g, returnLongestOnly = TRUE)
}
\arguments{
\item{transcripts}{GRanges object with ONLY exon annotations (no gene, transcript, CDS etc.) with all transcripts for orf retrevial}

\item{BSgenome}{BSgenome object}

\item{returnLongestOnly}{TRUE - currently only returns longest ORF}
}
\value{
data.frame with longest orf details
}
\description{
Get open reading frames for transcripts
}
\author{
Beth Signal
}
