% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{irfResults}
\alias{irfResults}
\alias{irfResults,irfDataSet-method}
\title{Method irfResults}
\usage{
irfResults(irfDataSet)

\S4method{irfResults}{irfDataSet}(irfDataSet)
}
\arguments{
\item{irfDataSet}{irfDataSet generated from \code{readIRFDataSet()}}
}
\value{
differential splicing results data.frame
(originally from a whippet .diff file)
}
\description{
Method irfResults
}
\examples{
irfinder_file <- list.files(system.file("extdata","irf_small/",
package="GeneStructureTools"), full.names=TRUE)
irf <- readIrfDataSet(irfinder_file)
irf.results <- irfResults(irfDataSet)
}
\concept{irf data processing}
