% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickIrfinder.R
\name{irfTranscriptChangeSummary}
\alias{irfTranscriptChangeSummary}
\title{Compare open reading frames for RMATS differentially spliced events}
\usage{
irfTranscriptChangeSummary(
  irfDataSet,
  exons = NULL,
  BSgenome,
  intronMatchType = "exact",
  NMD = TRUE,
  exportGTF = NULL
)
}
\arguments{
\item{irfDataSet}{irfsDataSet generated from \code{readIrfDataSet()}}

\item{exons}{GRanges gtf annotation of exons}

\item{BSgenome}{BSGenome object containing the genome for the species analysed}

\item{intronMatchType}{what type of matching to perform in findIntronContainingTranscripts?}

\item{NMD}{Use NMD predictions?}

\item{exportGTF}{file name to export alternative isoform GTFs (default=NULL)}
}
\value{
data.frame containing significant IRFinder differential splicing data and ORF change summaries
}
\description{
Compare open reading frames for RMATS differentially spliced events
}
\examples{
gtf <- rtracklayer::import(system.file("extdata", "gencode.vM25.small.gtf", package = "GeneStructureTools"))
exons <- gtf[gtf$type == "exon"]
g <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10

irfinder_file <- list.files(system.file("extdata", "irf_small/", package = "GeneStructureTools"), full.names = TRUE)
irf <- readIrfDataSet(irfinder_file)
irf.filtered <- filterIrfEvents(irf, FDR = 0.01, psiDelta = 0.1)
irf_summary <- irfTranscriptChangeSummary(irf.filtered, exons, BSgenome = g)
}
\seealso{
Other IRFinder data processing: 
\code{\link{filterIrfEvents}()},
\code{\link{readIrfDataSet}()}
}
\author{
Beth Signal
}
\concept{IRFinder data processing}
