% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{junctions}
\alias{junctions}
\alias{junctions,whippetDataSet-method}
\title{Method junctions}
\usage{
junctions(whippetDataSet)

\S4method{junctions}{whippetDataSet}(whippetDataSet)
}
\description{
Method junctions
}
