% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{leafcutterIntronsToExons}
\alias{leafcutterIntronsToExons}
\title{Create exon ranges from leafcutter intron ranges}
\usage{
leafcutterIntronsToExons(clusterGRanges, junctions)
}
\arguments{
\item{clusterGRanges}{GRanges of intron clusters with a 'set' column}
}
\value{
GRanges object with intron clusters converted to exons
}
\description{
Create exon ranges from leafcutter intron ranges
}
\seealso{
Other leafcutter splicing isoform creation: 
\code{\link{alternativeIntronUsage}()},
\code{\link{makeNewLeafExonsUnanchored}()},
\code{\link{makeNewLeafExons}()},
\code{\link{readLeafcutterJunctions}()}
}
\author{
Beth Signal
}
\concept{leafcutter splicing isoform creation}
\keyword{internal}
