% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickAnalysis.R
\name{leafcutterTranscriptChangeSummary}
\alias{leafcutterTranscriptChangeSummary}
\title{Compare open reading frames for whippet differentially spliced events}
\usage{
leafcutterTranscriptChangeSummary(
  leafcutterEvents,
  exons,
  FDR = NA,
  combineGeneEvents = FALSE,
  BSgenome,
  NMD = FALSE,
  showProgressBar = TRUE,
  junctions = NULL,
  exportGTF = NULL
)
}
\arguments{
\item{leafcutterEvents}{data.frame containing information from the
per_intron_results.tab file output from leafcutter.}

\item{exons}{GRanges gtf annotation of exons}

\item{FDR}{minimum FDR for events. If left as default (NA), will use FDR=0.05. To stop FDR filtering set to 1.}

\item{combineGeneEvents}{combine clusters occurring in the same gene?
Currently not recommended.}

\item{BSgenome}{BSGenome object containing the genome for the species analysed}

\item{NMD}{Use NMD predictions? (Note: notNMD must be installed to use this feature)}

\item{showProgressBar}{show a progress bar of alternative isoform generation?}

\item{junctions}{junctions GRanges object from readLeafcutterJunctions()}

\item{exportGTF}{file name to export alternative isoform GTFs (default=NULL)}
}
\value{
data.frame containing significant whippet diff data and ORF change summaries
}
\description{
Compare open reading frames for whippet differentially spliced events
}
\examples{
leafcutterFiles <- list.files(system.file("extdata", "leaf_small/", package = "GeneStructureTools"), full.names = TRUE)
leafcutterIntrons <- read.delim(leafcutterFiles[grep("intron_results", leafcutterFiles)], stringsAsFactors = FALSE)
gtf <- rtracklayer::import(system.file("extdata", "gencode.vM25.small.gtf", package = "GeneStructureTools"))
exons <- gtf[gtf$type == "exon"]
g <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
leafcutterTranscriptChangeSummary(leafcutterEvents = leafcutterIntrons, exons = exons, BSgenome = g, NMD = FALSE)
}
\author{
Beth Signal
}
\concept{leafcutter data processing}
