% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longestORF.R
\name{longestORF}
\alias{longestORF}
\title{Annotate introns and exonic parts by overlaping exon biotype}
\usage{
longestORF(nt_sequence)
}
\arguments{
\item{nt_sequence}{nucleotide sequence(s) to have longest orfs predicted}
}
\value{
data.frame with sequence index, length, and translation frame of the longest ORF predicted
}
\description{
Annotate introns and exonic parts by overlaping exon biotype
}
\author{
Beth Signal
}
