% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGvisTracks.R
\name{makeGeneModel}
\alias{makeGeneModel}
\title{Convert GRanges gene model to data.frame for visualisation with Gviz}
\usage{
makeGeneModel(transcript)
}
\arguments{
\item{transcripts}{GRanges of gene model to be visualised}
}
\value{
data.frame of a gene model for visualisation
}
\description{
Convert GRanges gene model to data.frame for visualisation with Gviz
}
\author{
Beth Signal
}
