% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{makeNewLeafExons}
\alias{makeNewLeafExons}
\title{Create new exon ranges for a cryptic splice junction}
\usage{
makeNewLeafExons(altIntronLocs, junctionRanges, clusterExons, splice.type)
}
\arguments{
\item{clusterExons}{GRanges of annotated exons from the target gene}

\item{splice.type}{which end of the junciton is cryptic? 3 or 5.}

\item{junctionGRanges}{GRanges of splice junction (must be intron version)}
}
\value{
GRanges object with new exons
}
\description{
Create new exon ranges for a cryptic splice junction
}
\seealso{
Other leafcutter splicing isoform creation: 
\code{\link{alternativeIntronUsage}()},
\code{\link{leafcutterIntronsToExons}()},
\code{\link{makeNewLeafExonsUnanchored}()},
\code{\link{readLeafcutterJunctions}()}
}
\author{
Beth Signal
}
\concept{leafcutter splicing isoform creation}
\keyword{internal}
