% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafCutter.R
\name{makeNewLeafExonsUnanchored}
\alias{makeNewLeafExonsUnanchored}
\title{Create new exon ranges for a cryptic unachored splice junction}
\usage{
makeNewLeafExonsUnanchored(altIntronLocs, junctionRanges, clusterExons)
}
\arguments{
\item{clusterExons}{GRanges of annotated exons from the target gene}

\item{junctionGRanges}{GRanges of splice junction (must be intron version)}
}
\value{
GRanges object with new exons
}
\description{
Create new exon ranges for a cryptic unachored splice junction
}
\seealso{
Other leafcutter splicing isoform creation: \code{\link{alternativeIntronUsage}},
  \code{\link{leafcutterIntronsToExons}},
  \code{\link{makeNewLeafExons}},
  \code{\link{readLeafcutterJunctions}}
}
\author{
Beth Signal
}
\keyword{internal}
