% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafCutter.R
\name{matrix2combinations}
\alias{matrix2combinations}
\title{Get a list of all potential range combinations, covering the full length of the region.}
\usage{
matrix2combinations(mat)
}
\arguments{
\item{mat}{matrix of all ranges by all ranges, with 0/1 coded overlaps}
}
\value{
list of range combinations possible
}
\description{
Get a list of all potential range combinations, covering the full length of the region.
}
\author{
Beth Signal
}
\keyword{internal}
