% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntronRetention.R
\name{maxLocation}
\alias{maxLocation}
\title{Find the largest distance between two vectors of numbers
Helper function for get_orfs}
\usage{
maxLocation(start_site, stop_site, longest = 1)
}
\arguments{
\item{start_site}{vector of start sites - i.e Met amino acid positions}

\item{stop_site}{vector of stop sites - i.e Stop (*) amino acid positions}

\item{longest}{1}
}
\value{
sequential start site and end site with the greatest difference
}
\description{
Find the largest distance between two vectors of numbers
Helper function for get_orfs
}
\author{
Beth Signal
}
