% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTranscripts.R
\name{orfDiff}
\alias{orfDiff}
\title{Evaluate changes to ORFs caused by alternative splicing}
\usage{
orfDiff(orfsX, orfsY, filterNMD = TRUE, geneSimilarity = TRUE,
  compareUTR = TRUE, compareBy = "gene", allORFs = NULL)
}
\arguments{
\item{orfsX}{orf information for 'normal' transcripts. Generated by getOrfs()}

\item{orfsY}{orf information for 'alternative' transcripts. Generated by getOrfs()}

\item{filterNMD}{filter orf information for transcripts not targeted by nmd first?}

\item{geneSimilarity}{compare orf to all orfs in gene?}

\item{compareUTR}{compare UTRs?}

\item{compareBy}{compare by 'transcript' isoforms or by 'gene' groups}

\item{allORFs}{orf information for all transcripts for novel sequence comparisons.
Generated by getOrfs()}
}
\value{
data.frame with orf changes
}
\description{
Evaluate changes to ORFs caused by alternative splicing
}
\examples{
gtf <- rtracklayer::import(system.file("extdata","example_gtf.gtf",
package = "GeneStructureTools"))
gtf.exons <- gtf[gtf$type=="exon"]
g <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
orfsProteinCoding <- getOrfs(gtf.exons[gtf.exons$gene_name=="Prex2" &
gtf.exons$transcript_type=="protein_coding"], BSgenome = g)
orfsNMD <- getOrfs(gtf.exons[gtf.exons$gene_name=="Prex2" &
gtf.exons$transcript_type=="nonsense_mediated_decay"], BSgenome = g)
orfChange <- orfDiff(orfsProteinCoding, orfsNMD, filterNMD=FALSE)
whippetFiles <- list.files(system.file("extdata","whippet/",
package = "GeneStructureTools"), full.names = TRUE)
diffFiles <- whippetFiles[grep(".diff", whippetFiles)]
whippetDiffSplice <- readWhippetDIFFfiles(diffFiles)
whippetCoords <- formatWhippetEvents(whippetDiffSplice)
event.exonSkip <- whippetDiffSplice[which(whippetDiffSplice$type=="CE")[1],]
coords.exonSkip <- whippetCoords[whippetCoords$id \%in\% event.exonSkip$coord]
exons.exonSkip <- findExonContainingTranscripts(coords.exonSkip, gtf.exons,
variableWidth=0, findIntrons=FALSE, gtf.transcripts)
ExonSkippingTranscripts <- skipExonInTranscript(coords.exonSkip, exons.exonSkip, gtf.exons)
orfsSkipped <- getOrfs(ExonSkippingTranscripts[ExonSkippingTranscripts$set=="skipped_exon"],
BSgenome = g)
orfsIncluded <- getOrfs(ExonSkippingTranscripts[ExonSkippingTranscripts$set=="included_exon"],
BSgenome = g)
orfDiff(orfsSkipped, orfsIncluded, filterNMD=FALSE)
}
\author{
Beth Signal
}
