% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTranscripts.R
\name{orfSimilarity}
\alias{orfSimilarity}
\title{calculate percentage of orfB contained in orfA}
\usage{
orfSimilarity(orfA, orfB)
}
\arguments{
\item{orfA}{character string of ORF amino acid sequence}

\item{orfB}{character string of ORF amino acid sequence}
}
\value{
percentage of orfB contained in orfA
}
\description{
calculate percentage of orfB contained in orfA
}
\author{
Beth Signal
}
