% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{overlap2matrix}
\alias{overlap2matrix}
\title{Create a 0/1 matrix from a Ranges overlap}
\usage{
overlap2matrix(ol, maxn = 7)
}
\arguments{
\item{ol}{ranges overlaps from findOverlaps()}
}
\value{
matrix of all ranges by all ranges, with 0/1 coded overlaps
}
\description{
Create a 0/1 matrix from a Ranges overlap
}
\author{
Beth Signal
}
\keyword{internal}
