% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{readCounts}
\alias{readCounts}
\alias{readCounts,whippetDataSet-method}
\title{Method readCounts}
\usage{
readCounts(whippetDataSet)

\S4method{readCounts}{whippetDataSet}(whippetDataSet)
}
\arguments{
\item{whippetDataSet}{FUUUUUUCk}
}
\description{
Method readCounts
}
