% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickIrfinder.R
\name{readIrfDataSet}
\alias{readIrfDataSet}
\title{Import IRFinder results files as a irfDataSet}
\usage{
readIrfDataSet(filePath)
}
\arguments{
\item{filePath}{path to IRFinder differential splicing output file}
}
\value{
irfDataSet
}
\description{
Import IRFinder results files as a irfDataSet
}
\examples{
irfinder_file <- list.files(system.file("extdata", "irf_small/", package = "GeneStructureTools"), full.names = TRUE)
irf <- readIrfDataSet(irfinder_file)
}
\seealso{
Other IRFinder data processing: 
\code{\link{filterIrfEvents}()},
\code{\link{irfTranscriptChangeSummary}()}
}
\author{
Beth Signal
}
\concept{IRFinder data processing}
