% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafCutter.R
\name{readLeafcutterJunctions}
\alias{readLeafcutterJunctions}
\title{Create junction ranges from STAR junction files}
\usage{
readLeafcutterJunctions(junctionFiles, minReads = 10)
}
\arguments{
\item{junctionFiles}{list of files with STAR junctions}

\item{minReads}{minimum number of reads a junction requires to be kept}
}
\value{
GRanges object with junctions
}
\description{
Create junction ranges from STAR junction files
}
\seealso{
Other leafcutter splicing isoform creation: \code{\link{alternativeIntronUsage}},
  \code{\link{leafcutterIntronsToExons}},
  \code{\link{makeNewLeafExonsUnanchored}},
  \code{\link{makeNewLeafExons}}
}
\author{
Beth Signal
}
