% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickRmats.R
\name{readRmatsDataSet}
\alias{readRmatsDataSet}
\title{Import RMATS Turbo results files as a rmatsDataSet}
\usage{
readRmatsDataSet(filePath, type = "JC")
}
\arguments{
\item{filePath}{path to RMATS differential splicing output files}

\item{type}{type of counts to use. "JC" or "JCEC"}
}
\value{
rmatsDataSet
}
\description{
Import RMATS Turbo results files as a rmatsDataSet
}
\examples{
rmats_filePath <- system.file("extdata", "rmats_small/", package = "GeneStructureTools")
rds <- readRmatsDataSet(rmats_filePath)
}
\seealso{
Other rmats data processing: 
\code{\link{altIntronRmats}()},
\code{\link{altSpliceSiteRmats}()},
\code{\link{annotateEventCoords}()},
\code{\link{annotateOverlapRmats}()},
\code{\link{betweenNumbers}()},
\code{\link{duplicateReference}()},
\code{\link{exonsToIntrons}()},
\code{\link{extractEvent}()},
\code{\link{filterRmatsEvents}()},
\code{\link{reformatExons}()},
\code{\link{removeDuplicatePairs}()},
\code{\link{removeExonsBetween}()},
\code{\link{rmatsTranscriptChangeSummary}()},
\code{\link{skipExonByJunction}()},
\code{\link{splitLongExons}()}
}
\author{
Beth Signal
}
\concept{rmats data processing}
