% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWhippetFiles.R
\name{readWhippetDataSet}
\alias{readWhippetDataSet}
\title{Import whippet results files as a whippetDataSet}
\usage{
readWhippetDataSet(filePath = ".")
}
\arguments{
\item{filePath}{path to whippet output files}
}
\value{
whippetDataSet
}
\description{
Import whippet results files as a whippetDataSet
}
\examples{
whippetFiles <- system.file("extdata","whippet/",
package = "GeneStructureTools")
wds <- readWhippetDataSet(whippetFiles)
}
\author{
Beth Signal
}
