% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWhippetFiles.R
\name{readWhippetPSIfiles}
\alias{readWhippetPSIfiles}
\title{Read in a list of whippet .psi.gz files and format as a data.frame}
\usage{
readWhippetPSIfiles(files, attribute = "Total_Reads", maxNA = NA)
}
\arguments{
\item{files}{vector of *.psi.gz file names}
}
\value{
data.frame with junction counts for all files
}
\description{
Read in a list of whippet .psi.gz files and format as a data.frame
}
\author{
Beth Signal
}
