% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exonReformatting.R
\name{reformatExons}
\alias{reformatExons}
\title{reformat a exons GRanges with first/last annotations and force 'biotype' to 'type'}
\usage{
reformatExons(exons)
}
\arguments{
\item{exons}{reference exons GRanges}
}
\value{
reference exons GRanges
}
\description{
reformat a exons GRanges with first/last annotations and force 'biotype' to 'type'
}
\examples{
gtf <- rtracklayer::import(system.file("extdata", "gencode.vM25.small.gtf",
    package = "GeneStructureTools"
))
exons <- gtf[gtf$type == "exon"]
exons.reformat <- reformatExons(exons)
}
\seealso{
Other rmats data processing: 
\code{\link{altIntronRmats}()},
\code{\link{altSpliceSiteRmats}()},
\code{\link{annotateEventCoords}()},
\code{\link{annotateOverlapRmats}()},
\code{\link{betweenNumbers}()},
\code{\link{duplicateReference}()},
\code{\link{exonsToIntrons}()},
\code{\link{extractEvent}()},
\code{\link{filterRmatsEvents}()},
\code{\link{readRmatsDataSet}()},
\code{\link{removeDuplicatePairs}()},
\code{\link{removeExonsBetween}()},
\code{\link{rmatsTranscriptChangeSummary}()},
\code{\link{skipExonByJunction}()},
\code{\link{splitLongExons}()}
}
\author{
Beth Signal
}
\concept{rmats data processing}
