% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AltSplicingJunctionSupported.R
\name{removeDuplicateTranscripts}
\alias{removeDuplicateTranscripts}
\title{Remove transcript duplicates}
\usage{
removeDuplicateTranscripts(transcripts)
}
\arguments{
\item{transcripts}{GRanges object with transcript structures in exon form}
}
\value{
GRanges object with unique transcript structures in exon form
}
\description{
Removes Structural duplicates of transcripts in a GRanges object
}
\author{
Beth Signal
}
