% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExonSkipping.R
\name{removeExonInTranscript}
\alias{removeExonInTranscript}
\title{Remove a skipped exon from the transcripts it is contained in}
\usage{
removeExonInTranscript(exonRanges, equal_exons, gtf.exons, glueExons = TRUE)
}
\arguments{
\item{exonRanges}{GRanges object with ranges for exons}

\item{equal_exons}{data.frame generataed by findExonContainingTranscripts()}

\item{gtf.exons}{GRanges object made from a GTF with ONLY exon annotations (no gene, transcript, CDS etc.)}

\item{glueExons}{Join together exons that are not seperated by exons?}
}
\value{
GRanges with transcripts skipping exons
}
\description{
Remove a skipped exon from the transcripts it is contained in
}
\author{
Beth Signal
}
