% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafCutter.R
\name{removeSameExon}
\alias{removeSameExon}
\title{Remove exon duplicates}
\usage{
removeSameExon(exons)
}
\arguments{
\item{exons}{GRanges object with exons}
}
\value{
GRanges object with unique exons
}
\description{
Removes structural duplicates of exons in a GRanges object
}
\examples{
gtf <- rtracklayer::import(system.file("extdata","example_gtf.gtf",
package = "GeneStructureTools"))
gtf.exons <- gtf[gtf$type=="exon"]
gtf.exons.duplicated <- c(gtf.exons[1:4], gtf.exons[1:4])
length(gtf.exons.duplicated)
gtf.exons.deduplicated <- removeSameExon(gtf.exons.duplicated)
length(gtf.exons.deduplicated)
}
\author{
Beth Signal
}
