% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exonReformatting.R
\name{removeVersion}
\alias{removeVersion}
\title{Remove version number from ensembl gene/transcript ids}
\usage{
removeVersion(ids)
}
\arguments{
\item{ids}{vector of ensembl ids}
}
\value{
vector of ensembl ids without the version number
}
\description{
Remove version number from ensembl gene/transcript ids
}
\examples{
removeVersion("ENSMUSG00000001017.15")
}
\author{
Beth Signal
}
