% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickRmats.R
\name{rmatsTranscriptChangeSummary}
\alias{rmatsTranscriptChangeSummary}
\title{Compare open reading frames for RMATS differentially spliced events}
\usage{
rmatsTranscriptChangeSummary(
  rmatsDataSet,
  exons = NULL,
  eventTypes = "all",
  BSgenome,
  NMD = TRUE,
  exportGTF = NULL
)
}
\arguments{
\item{rmatsDataSet}{rmatsDataSet generated from \code{readRmatsDataSet()}}

\item{exons}{GRanges gtf annotation of exons}

\item{eventTypes}{which event type to filter for? default="all"}

\item{BSgenome}{BSGenome object containing the genome for the species analysed}

\item{NMD}{Use NMD predictions? (Note: notNMD must be installed to use this feature)}

\item{exportGTF}{file name to export alternative isoform GTFs (default=NULL)}
}
\value{
data.frame containing significant RMATS differential splicing data and ORF change summaries
}
\description{
Compare open reading frames for RMATS differentially spliced events
}
\examples{
gtf <- rtracklayer::import(system.file("extdata", "gencode.vM25.small.gtf", package = "GeneStructureTools"))
exons <- gtf[gtf$type == "exon"]
g <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10

rmats_directory <- system.file("extdata", "rmats_small/", package = "GeneStructureTools")
rds <- readRmatsDataSet(rmats_directory)
rds.filtered <- filterRmatsEvents(rds, FDR = 0.01, psiDelta = 0.1)
rmats_summary <- rmatsTranscriptChangeSummary(rmatsDataSet = rds.filtered, exons, BSgenome = g)
}
\seealso{
Other rmats data processing: 
\code{\link{altIntronRmats}()},
\code{\link{altSpliceSiteRmats}()},
\code{\link{annotateEventCoords}()},
\code{\link{annotateOverlapRmats}()},
\code{\link{betweenNumbers}()},
\code{\link{duplicateReference}()},
\code{\link{exonsToIntrons}()},
\code{\link{extractEvent}()},
\code{\link{filterRmatsEvents}()},
\code{\link{readRmatsDataSet}()},
\code{\link{reformatExons}()},
\code{\link{removeDuplicatePairs}()},
\code{\link{removeExonsBetween}()},
\code{\link{skipExonByJunction}()},
\code{\link{splitLongExons}()}
}
\author{
Beth Signal
}
\concept{rmats data processing}
