% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{splitLongExons}
\alias{splitLongExons}
\title{Split long exons in two if they overlap an event}
\usage{
splitLongExons(betweenExons, gtfTranscripts)
}
\arguments{
\item{betweenExons}{data.frame with related differential splicing event ids and reference transcript_ids}

\item{gtfTranscripts}{Granges with altered transcript structures}
}
\value{
Granges with altered transcript structures
}
\description{
Split long exons in two if they overlap an event
}
\seealso{
Other rmats data processing: 
\code{\link{altIntronRmats}()},
\code{\link{altSpliceSiteRmats}()},
\code{\link{annotateEventCoords}()},
\code{\link{annotateOverlapRmats}()},
\code{\link{betweenNumbers}()},
\code{\link{duplicateReference}()},
\code{\link{exonsToIntrons}()},
\code{\link{extractEvent}()},
\code{\link{filterRmatsEvents}()},
\code{\link{readRmatsDataSet}()},
\code{\link{reformatExons}()},
\code{\link{removeDuplicatePairs}()},
\code{\link{removeExonsBetween}()},
\code{\link{rmatsTranscriptChangeSummary}()},
\code{\link{skipExonByJunction}()}
}
\author{
Beth Signal
}
\concept{rmats data processing}
\keyword{internal}
