% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverlapFunctions.R
\name{summariseExonTypes}
\alias{summariseExonTypes}
\title{Summarise exon biotypes to broader categories}
\usage{
summariseExonTypes(types)
}
\arguments{
\item{types}{vector of exon biotypes}
}
\value{
vector of broader exon biotypes
}
\description{
Summarise exon biotypes to broader categories
}
\author{
Beth Signal
}
