% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTR2UTR53.R
\name{UTR2UTR53}
\alias{UTR2UTR53}
\title{Annotate UTRs from Gencode GTF as 5' or 3'}
\usage{
UTR2UTR53(gtf)
}
\arguments{
\item{gtf}{GRanges object of the GTF}
}
\value{
gtf annotation GRanges object
}
\description{
Annotate UTRs from Gencode GTF as 5' or 3'
}
\examples{
gtf_file <- system.file("extdata","gencode.vM14.annotation.small.gtf",
package = "GeneStructureTools")
gtf <- rtracklayer::import(gtf_file)
gtf <- UTR2UTR53(gtf)
table(gtf$type)
}
\author{
Beth Signal
}
