% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickAnalysis.R
\name{whippetTranscriptChangeSummary}
\alias{whippetTranscriptChangeSummary}
\title{Compare open reading frames for whippet differentially spliced events}
\usage{
whippetTranscriptChangeSummary(significantEvents, whippetCoords,
  eventTypes = "all", gtf.exons, gtf.transcripts, BSgenome, jncCoords,
  NMD = FALSE)
}
\arguments{
\item{significantEvents}{data.frame containing signficant whippet diff data.}

\item{whippetCoords}{whippet coordinates for signficant whippet diff data.
Generated by formatWhippetEvents().}

\item{eventTypes}{which event type to filter for? default = "all"}

\item{gtf.exons}{GRanges gtf annotation of exons}

\item{gtf.transcripts}{GRanges gtf annotation of transcripts}

\item{BSgenome}{BSGenome object containing the genome for the species analysed}

\item{jncCoords}{whippet junction locations. Only needed if event types include AA/AD/AF/AL.}

\item{NMD}{Use NMD predictions? (Note: notNMD must be installed to use this feature)}
}
\value{
data.frame containing signficant whippet diff data and ORF change summaries
}
\description{
Compare open reading frames for whippet differentially spliced events
}
\examples{
whippetFiles <- list.files(system.file("extdata","whippet/",
package = "GeneStructureTools"), full.names = TRUE)
diffFiles <- whippetFiles[grep(".diff", whippetFiles)]
whippetDiffSplice <- readWhippetDIFFfiles(diffFiles)
whippetCoords <- formatWhippetEvents(whippetDiffSplice)
jncFiles <- whippetFiles[grep(".jnc", whippetFiles)]
jncCoords <- readWhippetJNCfiles(jncFiles)
jncCoords <- formatJunctions(jncCoords)
gtf <- rtracklayer::import(system.file("extdata","example_gtf.gtf",
package = "GeneStructureTools"))
gtf.exons <- gtf[gtf$type=="exon"]
gtf.transcripts <- gtf[gtf$type=="transcript"]
g <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
whippetTranscriptChangeSummary(significantEvents = whippetDiffSplice, whippetCoords,
eventTypes = "all",gtf.exons,gtf.transcripts,BSgenome = g,jncCoords = jncCoords, NMD=FALSE)
}
\author{
Beth Signal
}
