% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntronRetention.R
\name{addIntronInTranscript}
\alias{addIntronInTranscript}
\title{Add a retained intron to the transcripts it is skipped by}
\usage{
addIntronInTranscript(whippetDataSet, flankingExons, exons, glueExons = TRUE,
  match = "exact")
}
\arguments{
\item{whippetDataSet}{whippetDataSet generated from \code{readWhippetDataSet()}}

\item{flankingExons}{data.frame generataed by findIntronContainingTranscripts()}

\item{exons}{GRanges object made from a GTF with ONLY exon annotations
(no gene, transcript, CDS etc.)}

\item{glueExons}{Join together exons that are not seperated by introns?}

\item{match}{what type of match replacement should be done?
exact: exact matches to the intron only
retain: keep non-exact intron match coordinates in spliced sets, and retain them in retained sets
replace: replace non-exact intron match coordinates with event coordinates in spliced sets,
and retain in retained sets}
}
\value{
GRanges with transcripts containing retained introns
}
\description{
Add a retained intron to the transcripts it is skipped by
}
\examples{
whippetFiles <- system.file("extdata","whippet/",
package = "GeneStructureTools")
wds <- readWhippetDataSet(whippetFiles)
wds <- filterWhippetEvents(wds)

gtf <- rtracklayer::import(system.file("extdata","example_gtf.gtf",
package = "GeneStructureTools"))
exons <- gtf[gtf$type=="exon"]
g <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10

wds.intronRetention <- filterWhippetEvents(wds, eventTypes="RI")
exons.intronRetention <- findIntronContainingTranscripts(wds.intronRetention, exons)
IntronRetentionTranscripts <- addIntronInTranscript(wds.intronRetention,
exons.intronRetention, exons)
}
\author{
Beth Signal
}
