% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{coordinates}
\alias{coordinates}
\alias{coordinates,whippetDataSet-method}
\title{Method coordinates}
\usage{
coordinates(whippetDataSet)

\S4method{coordinates}{whippetDataSet}(whippetDataSet)
}
\arguments{
\item{whippetDataSet}{whippetDataSet generated from \code{readWhippetDataSet()}}
}
\value{
whippet splicing event coordinates as a GRanges object
}
\description{
Method coordinates
}
\examples{
whippetFiles <- system.file("extdata", "whippet_small/",
    package = "GeneStructureTools"
)
wds <- readWhippetDataSet(whippetFiles)

coordinates <- coordinates(wds)
}
\seealso{
Other whippet data processing: 
\code{\link{diffSplicingResults}()},
\code{\link{filterWhippetEvents}()},
\code{\link{formatWhippetEvents}()},
\code{\link{junctions}()},
\code{\link{readCounts}()},
\code{\link{readWhippetDIFFfiles}()},
\code{\link{readWhippetDataSet}()},
\code{\link{readWhippetJNCfiles}()},
\code{\link{readWhippetPSIfiles}()},
\code{\link{whippetTranscriptChangeSummary}()}
}
\concept{whippet data processing}
