% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{diffSplicingResults}
\alias{diffSplicingResults}
\alias{diffSplicingResults,whippetDataSet-method}
\title{Method diffSplicingResults}
\usage{
diffSplicingResults(whippetDataSet)

\S4method{diffSplicingResults}{whippetDataSet}(whippetDataSet)
}
\arguments{
\item{whippetDataSet}{whippetDataSet generated from \code{readWhippetDataSet()}}
}
\value{
differential splicing results data.frame
(originally from a whippet .diff file)
}
\description{
Method diffSplicingResults
}
\examples{
whippetFiles <- system.file("extdata","whippet/",
package = "GeneStructureTools")
wds <- readWhippetDataSet(whippetFiles)

diffSplicingResults <- diffSplicingResults(wds)
}
