% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exonReformatting.R
\name{exonsToTranscripts}
\alias{exonsToTranscripts}
\title{Convert an exon-level gtf annotation to a transcript-level gtf annotation}
\usage{
exonsToTranscripts(exons)
}
\arguments{
\item{exons}{GRanges object with exons}
}
\value{
GRanges object with transcripts
}
\description{
Convert an exon-level gtf annotation to a transcript-level gtf annotation
}
\examples{
gtf <- rtracklayer::import(system.file("extdata", "gencode.vM25.small.gtf",
    package = "GeneStructureTools"
))
exons <- gtf[gtf$type == "exon"]
transcritps <- exonsToTranscripts(exons)
}
\seealso{
Other gtf manipulation: 
\code{\link{UTR2UTR53}()},
\code{\link{removeDuplicateTranscripts}()},
\code{\link{removeSameExon}()},
\code{\link{reorderExonNumbers}()}
}
\author{
Beth Signal
}
\concept{gtf manipulation}
