% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverlapFunctions.R
\name{filterGtfOverlap}
\alias{filterGtfOverlap}
\title{Filter a GTF overlap to remove exons when exon is annotated as a CDS/UTR}
\usage{
filterGtfOverlap(gtf_from)
}
\arguments{
\item{gtf_from}{GRanges object of the GTF produced from an overlap}
}
\value{
GRanges object of the GTF with redundant exons removed
}
\description{
Filter a GTF overlap to remove exons when exon is annotated as a CDS/UTR
}
\author{
Beth Signal
}
