% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverlapFunctions.R
\name{findDEXexonType}
\alias{findDEXexonType}
\title{Find a DEXSeq exons' biotype}
\usage{
findDEXexonType(DEX_exon_id, dex_gtf, gtf, set = "overlap")
}
\arguments{
\item{DEX_exon_id}{vector of DEXSeq exon ids}

\item{dex_gtf}{GRanges object of the DEXSeq formatted gtf}

\item{gtf}{GRanges object of the GTF annotated with exon biotypes - i.e. exon, CDS, UTR}

\item{set}{which overlapping set of exon biotypes to return - to, from, and/or overlap}
}
\value{
overlaping types
}
\description{
Find a DEXSeq exons' biotype
}
\author{
Beth Signal
}
