% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExonSkipping.R
\name{findExonContainingTranscripts}
\alias{findExonContainingTranscripts}
\title{Given the location of a whole retained exon, find transcripts which can splice out this exon}
\usage{
findExonContainingTranscripts(exonRanges, gtf.exons)
}
\arguments{
\item{exonRanges}{GRanges object with ranges for exons}

\item{gtf.exons}{GRanges object made from a GTF with ONLY exon annotations (no gene, transcript, CDS etc.)}
}
\value{
data.frame with all overlapping exons
}
\description{
Given the location of a whole retained exon, find transcripts which can splice out this exon
}
\author{
Beth Signal
}
