% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AltSplicingJunctionSupported.R
\name{findJunctionPairs}
\alias{findJunctionPairs}
\title{Find alternative junctions for Whippet alternative splicing events}
\usage{
findJunctionPairs(exonRanges, jncRanges, type = NA)
}
\arguments{
\item{exonRanges}{GRanges object with Whippet event coordinates}

\item{jncRanges}{GRanges object with Whippet junctions. Generated by formatJunctions()}

\item{type}{type of Whippet event (AA/AD/AF/AL). Note only one event type should be processed at a time.}
}
\value{
GRanges object with alternative junctions. Each event should have a set of X (for which the psi measurement is reported) junctions, and alternative Y junctions.
}
\description{
Find junctions that pair with each end of an AA (alt. acceptor) or AD (alt. donor) whippet range
Find junctions that pair with the upsteam/downstream exon of an AF (alt. first exon) or an AL (alt. last exon)
}
\author{
Beth Signal
}
