% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{junctions}
\alias{junctions}
\alias{junctions,whippetDataSet-method}
\title{Method junctions}
\usage{
junctions(whippetDataSet)

\S4method{junctions}{whippetDataSet}(whippetDataSet)
}
\arguments{
\item{whippetDataSet}{whippetDataSet generated from \code{readWhippetDataSet()}}
}
\value{
junctions GRanges object
(originally from a whippet .jnc file)
}
\description{
Method junctions
}
\examples{
whippetFiles <- system.file("extdata", "whippet_small/",
    package = "GeneStructureTools"
)
wds <- readWhippetDataSet(whippetFiles)

junctions <- junctions(wds)
}
\seealso{
Other whippet data processing: 
\code{\link{coordinates}()},
\code{\link{diffSplicingResults}()},
\code{\link{filterWhippetEvents}()},
\code{\link{formatWhippetEvents}()},
\code{\link{readCounts}()},
\code{\link{readWhippetDIFFfiles}()},
\code{\link{readWhippetDataSet}()},
\code{\link{readWhippetJNCfiles}()},
\code{\link{readWhippetPSIfiles}()},
\code{\link{whippetTranscriptChangeSummary}()}
}
\concept{whippet data processing}
