% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTranscripts.R
\name{orfDiff}
\alias{orfDiff}
\title{Evaluate changes to ORFs caused by alternative splicing}
\usage{
orfDiff(orfsX, orfsY, filterNMD = TRUE, geneSimilarity = TRUE,
  compareUTR = TRUE)
}
\arguments{
\item{orfsX}{orf information for 'normal' transcripts. Generated by getOrfs()}

\item{orfsY}{orf information for 'alternative' transcripts. Generated by getOrfs()}

\item{filterNMD}{filter orf information for transcripts not targeted by nmd first?}

\item{geneSimilarity}{compare orf to all orfs in gene?}
}
\value{
data.frame with orf changes
}
\description{
Evaluate changes to ORFs caused by alternative splicing
}
\author{
Beth Signal
}
