% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTranscripts.R
\name{orfDiff}
\alias{orfDiff}
\title{Evaluate changes to ORFs caused by alternative splicing}
\usage{
orfDiff(orfs_x, orfs_y, filter_nmd = TRUE, gene_similarity = TRUE)
}
\arguments{
\item{orfs_x}{orf information for 'normal' transcripts. Generated by getOrfs()}

\item{orfs_y}{orf information for 'alternative' transcripts. Generated by getOrfs()}

\item{filter_nmd}{filter orf information for transcripts not targeted by nmd first?}

\item{gene_similarity}{compare orf to all orfs in gene?}
}
\value{
data.frame with orf changes
}
\description{
Evaluate changes to ORFs caused by alternative splicing
}
\author{
Beth Signal
}
