% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTranscripts.R
\name{orfDiff}
\alias{orfDiff}
\title{Evaluate changes to ORFs caused by alternative splicing}
\usage{
orfDiff(orfs_normal, orfs_alternative, transcript_matches)
}
\arguments{
\item{orfs_normal}{orf information for 'normal' transcripts. Generated by getOrfs()}

\item{orfs_alternative}{orf information for 'alternative' transcripts. Generated by getOrfs()}

\item{transcript_matches}{data.frame with all relationships to be tested}
}
\value{
data.frame with orf changes
}
\description{
Evaluate changes to ORFs caused by alternative splicing
}
\author{
Beth Signal
}
