% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTranscripts.R
\name{orfSimilarity}
\alias{orfSimilarity}
\title{calculate percentage of orf_b contained in orf_a}
\usage{
orfSimilarity(orf_a, orf_b)
}
\arguments{
\item{orf_a}{character string of ORF amino acid sequence}

\item{orf_b}{character string of ORF amino acid sequence}
}
\value{
percentage of orf_b contained in orf_a
}
\description{
calculate percentage of orf_b contained in orf_a
}
\author{
Beth Signal
}
