% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverlapFunctions.R
\name{overlapTypes}
\alias{overlapTypes}
\title{Annotate introns and exonic parts by overlaping exon biotype}
\usage{
overlapTypes(query_ranges, gtf, set = c("from", "to", "overlap"))
}
\arguments{
\item{query_ranges}{GRanges object of the query regions}

\item{gtf}{GRanges object of the GTF annotated with exon biotypes - i.e. exon, CDS, UTR}

\item{set}{which overlapping set of exon biotypes to return - to, from, and/or overlap}
}
\value{
overlaping types in a data.frame
}
\description{
Annotate introns and exonic parts by overlaping exon biotype
}
\author{
Beth Signal
}
