% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWhippetFiles.R
\name{readWhippetJNCfiles}
\alias{readWhippetJNCfiles}
\title{Read in a list of whippet .jnc.gz files and format as a data.frame}
\usage{
readWhippetJNCfiles(files)
}
\arguments{
\item{files}{vector of *.jnc.gz file names}
}
\value{
data.frame with junction counts for all files
}
\description{
Read in a list of whippet .jnc.gz files and format as a data.frame
}
\author{
Beth Signal
}
