% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AltSplicingJunctionSupported.R
\name{removeDuplicateTranscripts}
\alias{removeDuplicateTranscripts}
\title{Remove transcript duplicates}
\usage{
removeDuplicateTranscripts(transcripts)
}
\arguments{
\item{transcripts}{GRanges object with transcript structures in exon form}
}
\value{
GRanges object with unique transcript structures in exon form
}
\description{
Removes Structural duplicates of transcripts in a GRanges object
Note that duplicates must have different transcript ids.
}
\examples{
gtf <- rtracklayer::import(system.file("extdata","example_gtf.gtf",
package = "GeneStructureTools"))
gtf.exons <- gtf[gtf$type=="exon"]
gtf.exons.altName <- gtf.exons
gtf.exons.altName$transcript_id <- paste(gtf.exons.altName$transcript_id, "duplicated", sep="_")
gtf.exons.duplicated <- c(gtf.exons, gtf.exons.altName)
length(gtf.exons.duplicated)
gtf.exons.deduplicated <- removeDuplicateTranscripts(gtf.exons.duplicated)
length(gtf.exons.deduplicated)
}
\author{
Beth Signal
}
