% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AltSplicingJunctionSupported.R
\name{replaceJunction}
\alias{replaceJunction}
\title{Find transcripts containing/overlapping junctions and replace them with alternative junctions}
\usage{
replaceJunction(junctionPairs, exonRanges, gtf.exons, type = NA)
}
\arguments{
\item{junctionPairs}{GRanges object with alternative Whippet junctions. Generated by findJunctionPairs()}

\item{exonRanges}{GRanges object with Whippet event coordinates}

\item{type}{type of Whippet event (AA/AD/AF/AL). Note only one event type should be processed at a time.}
}
\value{
GRanges object with transcripts containing alternative junctions.
}
\description{
Find transcripts containing/overlapping junctions and replace them with alternative junctions
}
\author{
Beth Signal
}
