% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExonSkipping.R
\name{skipExonInTranscript}
\alias{skipExonInTranscript}
\title{Remove and iclude a skipped exon from the transcripts it overlaps}
\usage{
skipExonInTranscript(eventCoords, skippedExons, gtf.exons, glueExons = TRUE,
  match = "exact")
}
\arguments{
\item{eventCoords}{GRanges object with ranges for exons}

\item{skippedExons}{data.frame generataed by findExonContainingTranscripts()}

\item{gtf.exons}{GRanges object made from a GTF with ONLY exon annotations
(no gene, transcript, CDS etc.)}

\item{glueExons}{Join together exons that are not seperated by exons?}

\item{match}{what type of match replacement should be done?
exact: exact matches to the skipped event only, also removes any intron overlaps
skip: keep non-exact exon match coordinates in included sets, and skip them in skipped sets
replace: replace non-exact exon match coordinates with event coordinates in included sets,
and skip them in skipped sets}
}
\value{
GRanges with transcripts skipping exons
}
\description{
Remove and iclude a skipped exon from the transcripts it overlaps
}
\examples{
whippetFiles <- list.files(system.file("extdata","whippet/",
package = "GeneStructureTools"), full.names = TRUE)
diffFiles <- whippetFiles[grep(".diff", whippetFiles)]
whippetDiffSplice <- readWhippetDIFFfiles(diffFiles)
whippetCoords <- formatWhippetEvents(whippetDiffSplice)
gtf <- rtracklayer::import(system.file("extdata","example_gtf.gtf",
package = "GeneStructureTools"))
gtf.exons <- gtf[gtf$type=="exon"]
gtf.transcripts <- gtf[gtf$type=="transcript"]
event.exonSkip <- whippetDiffSplice[which(whippetDiffSplice$type=="CE")[1],]
coords.exonSkip <- whippetCoords[whippetCoords$id \%in\% event.exonSkip$coord]
exons.exonSkip <- findExonContainingTranscripts(coords.exonSkip, gtf.exons,
variableWidth=0, findIntrons=FALSE, gtf.transcripts)
ExonSkippingTranscripts <- skipExonInTranscript(coords.exonSkip, exons.exonSkip, gtf.exons)
}
\author{
Beth Signal
}
