% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverlapFunctions.R
\name{summariseExonTypes}
\alias{summariseExonTypes}
\title{Summarise exon biotypes to broader categories}
\usage{
summariseExonTypes(types)
}
\arguments{
\item{types}{vector of exon biotypes}
}
\value{
vector of broader exon biotypes
}
\description{
Summarise exon biotypes to broader categories
}
\examples{
gtfFile <- system.file("extdata","gencode.vM14.annotation.small.gtf",
package = "GeneStructureTools")
dexGtfFile <- system.file("extdata","gencode.vM14.annotation.dexseq.small.gtf",
package = "GeneStructureTools")

gtf <- rtracklayer::import(gtfFile)
gtf <- UTR2UTR53(gtf)
dexGtf <- rtracklayer::import(dexGtfFile)

DEXSeqResultsFile <- system.file("extdata","dexseq_results.small.txt",
package = "GeneStructureTools")
DEXSeqResults <- read.table(DEXSeqResultsFile, sep="\\t")

types <- findDEXexonType(rownames(DEXSeqResults), dexGtf, gtf)
summarisedTypes <- summariseExonTypes(types)
table(types, summarisedTypes)
}
\author{
Beth Signal
}
