% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickAnalysis.R
\name{transcriptChangeSummary}
\alias{transcriptChangeSummary}
\title{Compare open reading frames for two sets of paired transcripts}
\usage{
transcriptChangeSummary(transcriptsX, transcriptsY, g, NMD = FALSE,
  orfPrediction = "all_frames")
}
\arguments{
\item{transcriptsX}{GRanges object with exon annotations for all transcripts to be compared for the 'normal' condition}

\item{transcriptsY}{GRanges object with exon annotations for all transcripts to be compared for the 'alternative' condition}

\item{g}{BSGenome object containing the genome for the species analysed}

\item{NMD}{Use NMD predictions? (Note: notNMD must be installed to use this feature)}

\item{orfPrediction}{What type of orf predictions to return. default: "all_frames"}
}
\value{
Summarised ORF changes data.frame
}
\description{
Compare open reading frames for two sets of paired transcripts
}
\author{
Beth Signal
}
