% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickWhippet.R
\name{whippetTranscriptChangeSummary}
\alias{whippetTranscriptChangeSummary}
\title{Compare open reading frames for whippet differentially spliced events}
\usage{
whippetTranscriptChangeSummary(
  whippetDataSet,
  unfilteredWDS,
  BSgenome,
  eventTypes = "all",
  exons,
  NMD = FALSE,
  exportGTF = NULL,
  selectLongest = 1
)
}
\arguments{
\item{whippetDataSet}{whippetDataSet generated from \code{readWhippetDataSet()}}

\item{unfilteredWDS}{unfiltered whippetDataSet generated from \code{readWhippetDataSet()}
Note that this should contain ALL TS/TE events, regardless of significance, as these are used to construct the entire exon range.}

\item{BSgenome}{BSGenome object containing the genome for the species analysed}

\item{eventTypes}{which event type to filter for? default = "all"}

\item{exons}{GRanges gtf annotation of exons}

\item{NMD}{Use NMD predictions? (Note: notNMD must be installed to use this feature)}

\item{exportGTF}{file name to export alternative isoform GTFs (default=NULL)}

\item{selectLongest}{passed to getORFs()}
}
\value{
data.frame containing signficant whippet diff data and ORF change summaries
}
\description{
Compare open reading frames for whippet differentially spliced events
}
\examples{
gtf <- rtracklayer::import(system.file("extdata", "gencode.vM25.small.gtf", package = "GeneStructureTools"))
exons <- gtf[gtf$type == "exon"]
g <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10

whippetFiles <- system.file("extdata", "whippet_small/",
    package = "GeneStructureTools"
)
wds.all <- readWhippetDataSet(whippetFiles)
wds.signif <- filterWhippetEvents(wds.all)
whippetTranscriptChangeSummary(whippetDataSet = wds.signif, unfilteredWDS = wds.all, exons = exons, BSgenome = g)
}
\seealso{
Other whippet data processing: 
\code{\link{coordinates}()},
\code{\link{diffSplicingResults}()},
\code{\link{filterWhippetEvents}()},
\code{\link{formatWhippetEvents}()},
\code{\link{junctions}()},
\code{\link{readCounts}()},
\code{\link{readWhippetDIFFfiles}()},
\code{\link{readWhippetDataSet}()},
\code{\link{readWhippetJNCfiles}()},
\code{\link{readWhippetPSIfiles}()}
}
\author{
Beth Signal
}
\concept{whippet data processing}
