% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickAnalysis.R
\name{whippetTranscriptChangeSummary}
\alias{whippetTranscriptChangeSummary}
\title{Compare open reading frames for whippet differentially spliced events}
\usage{
whippetTranscriptChangeSummary(significantEvents, whippetCoords,
  eventTypes = "all", gtf.exons, g, whippetJnc, NMD = FALSE)
}
\arguments{
\item{significantEvents}{data.frame containing signficant whippet diff data.}

\item{eventTypes}{which event type to filter for? default = "all"}

\item{gtf.exons}{GRanges gtf annotation of exons}

\item{g}{BSGenome object containing the genome for the species analysed}

\item{whippetJnc}{whippet junction locations. Only needed if event types include AA/AD/AF/AL.}

\item{NMD}{Use NMD predictions? (Note: notNMD must be installed to use this feature)}
}
\value{
data.frame containing signficant whippet diff data and ORF change summaries
}
\description{
Compare open reading frames for whippet differentially spliced events
}
\author{
Beth Signal
}
